requirements.txt
Django==4
django-cors-headers
djangorestframework
requests
pandas
python-dotenv
